requirements.txt
torch
torchvision
pillow
packaging
numpy
opencv-python